fact(tunnels): separate linux tunneling functionality
